// csp-angular19/src/app/models/lesson-learnt-model.ts
export class LessonLearntModel {
    id: number = 0;
    projecT_ID: string = '';
    description: string = '';
    publisheD_BY: string = '';
    publisheD_DATE: Date | null = null;
    categorY_OF_LESSON: string = '';
    procesS_AREA: string = '';
    createD_BY: string = '';
    createD_DATE: Date | null = null;
    updateD_BY: string = '';
    updateD_DATE: Date | null = null;
    isactive: boolean = true;
}